Refactor yk module in libcore
