+++ src/test/ui/yk_swt/collect-traces-threaded.rs
@@ -19,7 +19,8 @@ extern crate core;
 extern crate libc;
 extern crate test;
 
-use core::yk_swt::{start_tracing, stop_tracing, SirLoc};
+use core::yk::{SirLoc};
+use core::yk::swt::{start_tracing, stop_tracing};
 use std::thread;
 use test::black_box;
